Add custom exception for controller, service, repository

+++ backend/app/Services/UserService.php
@@ -6,10 +6,9 @@ use App\Exceptions\CustomException;
 use App\Interfaces\IUserService;
 use App\Models\User;
 use App\Repositories\IUserRepository;
-use Exception;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Pagination\LengthAwarePaginator;
-use Illuminate\Support\Facades\Log;
+
 
 class UserService implements IUserService
 {
@@ -24,9 +23,15 @@ class UserService implements IUserService
         return $this->userRepository->GetRoles();
     }
 
-    public function GetUserById(int $userId): ?User
+    /**
+     * @throws CustomException
+     */
+    public function GetUserDetailsById(int $userId): User
     {
-        return $this->userRepository->GetUserById($userId);
+        $result = $this->userRepository->GetUserDetailsById($userId);
+        if(!$result)
+            throw new CustomException('User not found. Please contact the administrator');
+        return $result;
     }
 
     public function SearchUsers(array $filters): ?LengthAwarePaginator
@@ -39,13 +44,25 @@ class UserService implements IUserService
         return $this->userRepository->AddUser($fields);
     }
 
+    /**
+     * @throws CustomException
+     */
     public function UpdateUser(array $fields): bool
     {
+        $user = $this->userRepository->GetUserById($fields['userId']);
+        if(!$user)
+            throw new CustomException('User not found!');
         return $this->userRepository->UpdateUser($fields);
     }
 
+    /**
+     * @throws CustomException
+     */
     public function DeleteUser(int $userId): bool
     {
+        $user = $this->userRepository->GetUserById($userId);
+        if(!$user)
+            throw new CustomException('User not found!');
         return $this->userRepository->DeleteUser($userId);
     }
